Bug 3065002 - Now displays number of rows for tables with more than 2,147,483,647 rows.
git-svn-id: https://schemaspy.svn.sourceforge.net/svnroot/schemaspy/trunk@599 e86a8a9b-1e1f-0410-8769-81452b7583c3

+++ src/net/sourceforge/schemaspy/model/View.java
@@ -68,8 +68,8 @@ public class View extends Table {
     }
 
     @Override
-    protected int fetchNumRows() {
-        return 0;
+    protected long fetchNumRows() {
+        return 0L;
     }
 
     /**
